FindVars {wrappedtools}	R Documentation

Find numeric index and names of column name patterns

Description

FindVars looks up columnames (by default for data-frame rawdata) based on parts of names, using regular expressions. exlusion rules may be specified as well.

Usage

FindVars(
  varnames,
  allnames = colnames(rawdata),
  exact = F,
  exclude = NA,
  casesensitive = T,
  fixed = F
)

Arguments

varnames
Vector of pattern to look for.

allnames
Vector of values to detect pattern in; by default, colnames(rawdata).

exact
Partial matching or exact only (adding ^ and $)?

exclude
Vector of pattern to exclude from found names.

casesensitive
Logical if case is respected in matching (default FALSE: a<>A)

fixed
Logical, treat pattern as regex?

Value

A list with index, names, backticked names, and symbols

Examples

FindVars(varnames = c('^c','g'),allnames = colnames(mtcars))
FindVars(varnames = c('^c','g'),allnames = colnames(mtcars),exclude='r')

[Package wrappedtools version 0.4.03 Index]