compare2numvars {wrappedtools}	R Documentation

Comparison for columns of numbers for 2 groups

Description

Comparison for columns of numbers for 2 groups

Usage

compare2numvars(
  data,
  testvars,
  groupvar,
  gaussian,
  round_p = 3,
  round_desc = 2,
  range = F,
  rangesep = " ",
  pretext = F,
  mark = F,
  n = F,
  .n = F
)

Arguments

data
name of dataset (tibble/data.frame)to analyze.

testvars
vector of column names.

groupvar
name of grouping variable, has to translate to 2 groups.

gaussian
logical specifying normal or ordinal values.

round_p
level for rounding p-value

round_desc
number of significant digits for rounding of descriptive stats

range
include min/max? qparam rangesep test between statistics and range or other elements

pretext
for function formatP

mark
for function formatP

n
create columns for n per group?

.n
add n to descriptive statistics?

[Package wrappedtools version 0.4.03 Index]