formatP {wrappedtools}	R Documentation

Re-format p-values

Description

formatP simplifies p-values by rounding to the maximum of p or a predefined level. Optionally < or = can be added, as well as symbols according to significance level.

Usage

formatP(pIn, ndigits = 3, text = T, pretext = F, mark = F, german_num = F)

Arguments

pIn
A numeric vector or matrix with p-values.

ndigits
Number of digits (default=3).

text
Should output be casted to character default=T)?

pretext
Should = or < be added before p (default=F)?

mark
Should significance level be added after p (default=F)?

german_num
change dot (default) to comma?

[Package wrappedtools version 0.4.06 Index]