meanse {wrappedtools}	R Documentation

Compute mean and se and put together with the +- symbol.

Description

Compute mean and se and put together with the +- symbol.

Usage

meanse(x, mult = 1, roundDig = 2, drop0 = F)

Arguments

x
Data for computation.

mult
multiplier for se, default 1, can be set to 2 or 1.96 to create confidence intervals

roundDig
Number of relevant digits for roundR.

drop0
Should trailing zeros be dropped?

[Package wrappedtools version 0.4.03 Index]