pwr_poisson {wrappedtools}	R Documentation

Computation of sample size for traits with Poisson distribution

Description

Computation of sample size for traits with Poisson distribution

Usage

pwr_poisson(
  .lambda_ctr,
  .lambda_ctr_range = 1,
  .lambda_delta,
  .lambda_delta_range = 1,
  .lambda_step = 0.5,
  .nrepl = 10^3,
  .n_min = 4,
  .label_ctr = "Lambda control",
  .label_verum = "Lambda verum"
)

Arguments

.lambda_ctr
expected count in control group.

.lambda_delta
Expected difference on counts for treatment.

.lambda_delta_range
Range for optimistic/pessimistic estimation for treatment. Within range, lambda is increased stepwise.

.lambda_step
defines stepwise increase.

.nrepl
Number of replications for simulation.

.n_min
Smallest sample size to start with.

.label_ctr
Label for plot.

.label_verum
Label for plot.

.lamba_ctr_range
Range for optimistic/pessimistic estimation for controls, resulting in three estimations.

[Package wrappedtools version 0.4.03 Index]