roundR {wrappedtools}	R Documentation

Automatic rounding to a reasonable length, based on largest number

Description

roundR returns a matrix of rounded numbers.

Usage

roundR(roundin, smooth = F, level = 2, textout = T, drop0 = F, .german = F)

Arguments

roundin
A vector or matrix of numbers.

smooth
A logical specifying if you want rounding before the dot (e.g. 12345 to 12300).

level
A number specifying number of relevant digits to keep.

textout
A logical if output is converted to text.

drop0
A logical if trailing zeros should be dropped

.german
A logical if german numbers should be reported

[Package wrappedtools version 0.4.03 Index]